Moves button inside keypad
